break if wp-cli is installed

+++ install-wp-cli.sh
@@ -1,5 +1,9 @@
 #!/usr/bin/env bash
 
+if [ -f /usr/local/bin/wp ]; then
+    echo "wp-cli is installed"
+    exit 2
+fi
 curl -O https://raw.githubusercontent.com/wp-cli/builds/gh-pages/phar/wp-cli.phar
 if [ -d /etc/php5 ]; then
     apt-get install php5-cli -y
